requirements.txt
# ── Resume parsing (Step 1) ───────────────────────────────────────────────────
PyMuPDF==1.24.5
openai>=1.0.0
python-dotenv==1.0.1

# ── Embeddings (Step 2) ───────────────────────────────────────────────────────
sentence-transformers>=3.3.0
torch>=2.4
faiss-cpu>=1.8.0
numpy>=1.24
scikit-learn>=1.4

# ── Job scraping (Step 3) ─────────────────────────────────────────────────────
psycopg[binary]>=3.1
psycopg-pool>=3.1
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.44
APScheduler>=3.10
playwright-stealth>=1.0
